refactor: 공통 모델 패키지명 수정
- domain -> common
