+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -8,7 +8,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import kitchenpos.domain.Price;
+import kitchenpos.common.Price;
 import kitchenpos.product.domain.Product;
 
 @Entity
