+++ src/test/java/kitchenpos/product/domain/NameTest.java
@@ -3,7 +3,7 @@ package kitchenpos.product.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-import kitchenpos.domain.Name;
+import kitchenpos.common.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
